+++ .eslintrc.js
@@ -16,7 +16,8 @@ module.exports = {
   },
   plugins: ['simple-import-sort'],
   rules: {
-    'simple-import-sort/sort': 'error',
+    'simple-import-sort/imports': 'error',
+    'simple-import-sort/exports': 'error',
     'import/order': 'off',
     'react/prop-types': 'off',
     'no-unused-vars': [