+++ packages/core/src/index.js
@@ -1,5 +1,5 @@
-export { default as useCorn } from './useCorn'
-export { default as useCornField } from './useCornField'
 export * from './fields'
 export * from './hooks'
 export * from './object'
+export { default as useCorn } from './useCorn'
+export { default as useCornField } from './useCornField'